fix(deepagent-code): isolate http api exerciser state
