+++ packages/deepagent-code/script/httpapi-exercise.ts
@@ -1,1 +1,2 @@
+await import("../test/server/httpapi-exercise/environment-bootstrap")
 await import("../test/server/httpapi-exercise/index")
