+++ packages/deepagent-code/test/server/httpapi-exercise/environment.ts
@@ -1,24 +1,16 @@
 import { Flag } from "@deepagent-code/core/flag/flag"
 import { Effect } from "effect"
-import path from "path"
+import {
+  exerciseConfigDirectory,
+  exerciseDataDirectory,
+  exerciseDatabasePath,
+  exerciseGlobalRoot,
+  preserveExerciseDatabase,
+  preserveExerciseGlobalRoot,
+} from "./environment-bootstrap"
 
-const preserveExerciseGlobalRoot = !!process.env.DEEPAGENT_CODE_HTTPAPI_EXERCISE_GLOBAL
-export const exerciseGlobalRoot =
-  process.env.DEEPAGENT_CODE_HTTPAPI_EXERCISE_GLOBAL ??
-  path.join(process.env.TMPDIR ?? "/tmp", `deepagent-code-httpapi-global-${process.pid}`)
-process.env.XDG_DATA_HOME = path.join(exerciseGlobalRoot, "data")
-process.env.XDG_CONFIG_HOME = path.join(exerciseGlobalRoot, "config")
-process.env.XDG_STATE_HOME = path.join(exerciseGlobalRoot, "state")
-process.env.XDG_CACHE_HOME = path.join(exerciseGlobalRoot, "cache")
-process.env.DEEPAGENT_CODE_DISABLE_SHARE = "true"
-export const exerciseConfigDirectory = path.join(exerciseGlobalRoot, "config", "deepagent-code")
-export const exerciseDataDirectory = path.join(exerciseGlobalRoot, "data", "deepagent-code")
+export { exerciseConfigDirectory, exerciseDataDirectory, exerciseDatabasePath, exerciseGlobalRoot }
 
-const preserveExerciseDatabase = !!process.env.DEEPAGENT_CODE_HTTPAPI_EXERCISE_DB
-export const exerciseDatabasePath =
-  process.env.DEEPAGENT_CODE_HTTPAPI_EXERCISE_DB ??
-  path.join(process.env.TMPDIR ?? "/tmp", `deepagent-code-httpapi-exercise-${process.pid}.db`)
-process.env.DEEPAGENT_CODE_DB = exerciseDatabasePath
 Flag.DEEPAGENT_CODE_DB = exerciseDatabasePath
 
 export const original = {
